Remove (the just moved) hub plugin (bloated)
I do not like/need the "alias git=hub" in general.

Especially in the way it was done here: it is function to set
`_ZSH_VCS_INFO_FORCE_GETDATA=1` to invalidate my custom vcs_info/git cache.
But this is a) not critical and b) is done more general by looking at
the executed command.

The extra stuff was needed because a function was used instead of an
alias.